[openMVG_Samples] Add a tutorial about 2View robust model estimation. #539 Add a demo sample to the OpenMVG_Sample demo directory:
It shows how to estimate a homography thanks to the OpenMVG robust estimation framework:
- MaxConsensus
- Ransac
- LMeds
- ACRansac

+++ src/openMVG/robust_estimation/robust_estimator_Ransac.hpp
@@ -35,7 +35,7 @@ typename Kernel::Model RANSAC(
   const Kernel &kernel,
   const Scorer &scorer,
   std::vector<size_t> *best_inliers = nullptr ,
-  double *best_score = nullptr , // Found number of inliers
+  size_t *best_score = nullptr , // Found number of inliers
   double outliers_probability = 1e-2)
 {
   assert(outliers_probability < 1.0);
